Finished the maximum flow algorithm.  It deserves an explanation: I looked at a couple of available algorithms like Ford-fulkerson algorithm, Edmonson-Karp and Dinic, but I didn't like them due to their naive usage of BFS, which leads to a lot of node visits.
I therefore choose to invert the capacities of the graph so that the capacity any G[u][v] = c becomes 1/c in G_inverted.
This allows me to use the shortest path method to find the path with most capacity in the first attempt, resulting in a significant
reduction of unassigned flow.
By updating G_inverted, with the residual capacity, I can keep using the shortest path, until the capacity is zero, whereby I remove the links
When the shortest path method returns 'No path' or when unassigned flow is zero, I exit the algorithm.

The method seems very efficient.